Update app source code
- Parse returned routes
- Get current time correctly
  - Departure.etd is now long type, was uint
- Make recv_body_buf a local var (was global)
  - Prevents segfault with muti-threading
- Add curl_global_init
  - Better thread safety
  - Don't init win32 socket lib
- Replace most printf calls with nxt_unit_req_log

+++ src/stop.h
@@ -21,7 +21,7 @@
 
 typedef struct Departure {
   char display_text[50];
-  unsigned int etd;
+  long etd;
 } Departure;
 
 typedef struct RouteDirection {
